requirements.txt
pyspark
delta-spark
requests